+++ Sources/CardPayments/CardClient.swift
@@ -40,7 +40,7 @@ public class CardClient: NSObject {
         analyticsService?.sendEvent("card-payments:3ds:started")
         Task {
             do {
-                let confirmPaymentRequest = try ConfirmPaymentSourceRequest(accessToken: config.accessToken, cardRequest: request)
+                let confirmPaymentRequest = try ConfirmPaymentSourceRequest(clientID: config.clientID, cardRequest: request)
                 let (result) = try await apiClient.fetch(request: confirmPaymentRequest)
                 
                 if let url: String = result.links?.first(where: { $0.rel == "payer-action" })?.href {
@@ -50,11 +50,7 @@ public class CardClient: NSObject {
                 } else {
                     analyticsService?.sendEvent("card-payments:3ds:confirm-payment-source:succeeded")
                     
-                    let cardResult = CardResult(
-                        orderID: result.id,
-                        status: result.status,
-                        paymentSource: result.paymentSource
-                    )
+                    let cardResult = CardResult(orderID: result.id, deepLinkURL: nil)
                     notifySuccess(for: cardResult)
                 }
             } catch let error as CoreSDKError {
@@ -88,7 +84,7 @@ public class CardClient: NSObject {
                     self?.analyticsService?.sendEvent("card-payments:3ds:challenge-presentation:failed")
                 }
             },
-            sessionDidComplete: { _, error in
+            sessionDidComplete: { url, error in
                 self.delegate?.cardThreeDSecureDidFinish(self)
                 if let error = error {
                     switch error {
@@ -100,31 +96,11 @@ public class CardClient: NSObject {
                         return
                     }
                 }
-                self.getOrderInfo(id: orderId)
+                
+                let cardResult = CardResult(orderID: orderId, deepLinkURL: url)
+                self.notifySuccess(for: cardResult)
             }
         )
-    }
-
-    private func getOrderInfo(id orderId: String) {
-        let getOrderInfoRequest = GetOrderInfoRequest(
-            orderID: orderId,
-            accessToken: config.accessToken
-        )
-        Task {
-            do {
-                let (result) = try await apiClient.fetch(request: getOrderInfoRequest)
-                let cardResult = CardResult(
-                    orderID: result.id,
-                    status: result.status,
-                    paymentSource: result.paymentSource
-                )
-                self.analyticsService?.sendEvent("card-payments:3ds:get-order-info:succeeded")
-                notifySuccess(for: cardResult)
-            } catch let error as CoreSDKError {
-                self.analyticsService?.sendEvent("card-payments:3ds:get-order-info:failed")
-                notifyFailure(with: error)
-            }
-        }
     }
 
     private func notifySuccess(for result: CardResult) {